requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
psycopg[binary]==3.2.10
